Added bitdepth rescaling (properly) to get_px_array and added re-change back to float type in creating MIPPYImage objects as a patch. MIPPYImage needs to be able to work with integer arrays inherently.

+++ MIPPY_new_struct/mippy/viewing.py
@@ -1030,7 +1030,7 @@ class MIPPYImage():
 		return
 
 	def construct_from_array(self,pixel_array):
-		self.px_float = pixel_array
+		self.px_float = pixel_array.astype(np.float64)
 		self.rangemax = np.amax(pixel_array)
 		self.rangemin = np.amin(pixel_array)
 #		print "Max",self.rangemax
